+++ tests/conftest.py
@@ -40,7 +40,7 @@ def geojson_point():
             "type": "Point",
             "coordinates": [13.397788148643649, 52.52958586909979],
         },
-        "properties": {"strname": "Unknown Road"},
+        "properties": {"str_name": "Unknown Road"},
     }